Message based card shuffler
Pivoted to a card shuffler because it is an easy concept to think about.
The card shuffler is made up of various services that communicate with
each other via Kafka messages.  Once a request is made to the API,
the deck process gets kicked off.  The deck process splits half the
deck to a top topic, and the other half to a bottom topic.  The top and
bottom topics each take 1-3 cards from their message, and send that small
packet to the shuffle topic.  The shuffle process waits until all cards
from the deck have been received.  If there are more shuffles to do,
it decrements the count and posts to the deck topic.  If all shuffles are
done, it posts to the output topic.  Even though all the processes are
technically sequential, the processors are immplemented in separate threads.