Add randomness per 'riffle', add to README description.
Because there was no delay between consuming from the top and bottom
topics and producing to the shuffled topic, the riffles weren't very
random.  Now, each time a message is produced to the shuffled topic,
the top and bottom hand sleep for 1 or 3 ms.  This slows the algorithm
a bit, but seems to produce a better riffle.

The README was a little sparse, so spruced it up with a better description
of what the code is doing.